Move the rest of training date related methods to TrainingDates and make them return TrainingDate
